+++ PageViews/PageViews.js
@@ -1,6 +1,7 @@
 import handle, {forwardCustomWithPrevent} from '@enact/core/handle';
-import kind from '@enact/core/kind';
 import EnactPropTypes from '@enact/core/internal/prop-types';
+import kind from '@enact/core/kind';
+import {cap} from '@enact/core/util';
 import {I18nContextDecorator} from '@enact/i18n/I18nDecorator';
 import SpotlightContainerDecorator, {spotlightDefaultClass} from '@enact/spotlight/SpotlightContainerDecorator';
 import Changeable from '@enact/ui/Changeable';
@@ -127,6 +128,15 @@ const PageViewsBase = kind({
 		onWillTransition: PropTypes.func,
 
 		/**
+		 * Specifies on which side (`'top'` or `'bottom'`) the page indicator appears.
+		 *
+		 * @type {('top'|'bottom')}
+		 * @default 'bottom'
+		 * @private
+		 */
+		pageIndicatorPosition: PropTypes.oneOf(['top', 'bottom']),
+
+		/**
 		 * Type of page indicator.
 		 *
 		 * There are two types:
@@ -159,6 +169,7 @@ const PageViewsBase = kind({
 
 	defaultProps: {
 		arranger: BasicArranger,
+		pageIndicatorPosition: 'bottom',
 		pageIndicatorType: 'dot'
 	},
 
@@ -202,7 +213,7 @@ const PageViewsBase = kind({
 	},
 
 	computed: {
-		className: ({fullContents, pageIndicatorType, styler}) => styler.append({fullContents}, pageIndicatorType),
+		className: ({fullContents, pageIndicatorPosition, pageIndicatorType, styler}) => styler.append({fullContents}, `indicator${cap(pageIndicatorPosition)}`, pageIndicatorType),
 		renderNextButton: ({css, onNextClick, index, totalIndex}) => {
 			const isNextButtonVisible = index < totalIndex - 1;
 
@@ -280,6 +291,7 @@ const PageViewsBase = kind({
 		componentRef,
 		fullContents,
 		index,
+		pageIndicatorPosition,
 		pageIndicatorType,
 		renderNextButton,
 		renderPrevButton,
@@ -300,7 +312,7 @@ const PageViewsBase = kind({
 
 		return (
 			<div role="region" aria-labelledby={`pageViews_index_${index}`} ref={componentRef} {...rest}>
-				{!fullContents && pageIndicatorType === 'dot' ? steps : null}
+				{!fullContents && pageIndicatorPosition === 'top' ? steps : null}
 				<Column aria-label={stepHintAriaLabel} className={css.contentsArea} id={`pageViews_index_${index}`} >
 					{fullContents ?
 						<>
@@ -315,11 +327,22 @@ const PageViewsBase = kind({
 						</Row>
 					}
 				</Column>
-				{!fullContents && pageIndicatorType === 'number' ? steps : null}
+				{!fullContents && pageIndicatorPosition === 'bottom' ? steps : null}
 			</div>
 		);
 	}
 });
+
+/**
+ * Sets the strategy used to automatically focus an element within the PageViews upon render.
+ * When set to 'none', focus is not set only on the first render.
+ *
+ * @name autoFocus
+ * @type {('default-element'|'last-focused'|'none'|String)}
+ * @memberof sandstone/PageViews.PageViews.prototype
+ * @default 'last-focused'
+ * @public
+ */
 
 const PageViewsDecorator = compose(
 	Changeable({prop: 'index'}),