+++ tests/screenshot/apps/components/PageViews.js
@@ -18,6 +18,26 @@ const PageComponents = [
 
 const BaseTests = [
 	{
+		component: <Panel><Header title="title of panel" subtitle="subtitle of panel" /><PageViews index={0} pageIndicatorPosition="top">{PageComponents[0]}</PageViews></Panel>,
+		wrapper: {full: true}
+	},
+	{
+		component: <Panel><Header title="title of panel" subtitle="subtitle of panel" /><PageViews index={0} pageIndicatorPosition="top">{PageComponents}</PageViews></Panel>,
+		wrapper: {full: true}
+	},
+	{
+		component: <Panel><Header title="title of panel" subtitle="subtitle of panel" /><PageViews index={1} pageIndicatorPosition="top">{PageComponents}</PageViews></Panel>,
+		wrapper: {full: true}
+	},
+	{
+		component: <Panel><Header title="title of panel" subtitle="subtitle of panel" /><PageViews index={2} pageIndicatorPosition="top">{PageComponents}</PageViews></Panel>,
+		wrapper: {full: true}
+	},
+	{
+		component: <Panel><Header title="title of panel" subtitle="subtitle of panel" /><PageViews fullContents index={2} pageIndicatorPosition="top">{PageComponents}</PageViews></Panel>,
+		wrapper: {full: true}
+	},
+	{
 		component: <Panel><Header title="title of panel" subtitle="subtitle of panel" /><PageViews index={0}>{PageComponents[0]}</PageViews></Panel>,
 		wrapper: {full: true}
 	},
@@ -55,6 +75,22 @@ const BaseTests = [
 	},
 	{
 		component: <Panel><Header title="title of panel" subtitle="subtitle of panel" /><PageViews pageIndicatorType="number" index={2}>{PageComponents}</PageViews></Panel>,
+		wrapper: {full: true}
+	},
+	{
+		component: <Panel><Header title="title of panel" subtitle="subtitle of panel" /><PageViews pageIndicatorType="number" pageIndicatorPosition="top" index={0}>{PageComponents[0]}</PageViews></Panel>,
+		wrapper: {full: true}
+	},
+	{
+		component: <Panel><Header title="title of panel" subtitle="subtitle of panel" /><PageViews pageIndicatorType="number" index={0} pageIndicatorPosition="top">{PageComponents}</PageViews></Panel>,
+		wrapper: {full: true}
+	},
+	{
+		component: <Panel><Header title="title of panel" subtitle="subtitle of panel" /><PageViews pageIndicatorType="number" index={1} pageIndicatorPosition="top">{PageComponents}</PageViews></Panel>,
+		wrapper: {full: true}
+	},
+	{
+		component: <Panel><Header title="title of panel" subtitle="subtitle of panel" /><PageViews pageIndicatorType="number" index={2} pageIndicatorPosition="top">{PageComponents}</PageViews></Panel>,
 		wrapper: {full: true}
 	}
 ];