+++ CHANGELOG.md
@@ -6,8 +6,13 @@ The following is a curated list of changes in the Enact sandstone module, newest
 
 ### Added
 
+- `sandstone/PageViews` prop `autoFocus` to set whether focus element automatically or not
 - `sandstone/PageViews.Page` and `sandstone/QuickGuidePanels.Panel` prop `aria-label`
 - `sandstone/QuickGuidePanels` props `closeButtonAriaLabel` and `onClose`
+
+### Fixed
+
+- `sandstone/PageViews` to not clip the shadow of navigation buttons when `fullContents` prop is `true`
 
 ## [3.0.0-alpha.2] - 2024-10-08